+++ src/onegov.wtfs/onegov/wtfs/fields/__init__.py
@@ -1,10 +1,8 @@
-from onegov.wtfs.fields.csv import CsvUploadField
 from onegov.wtfs.fields.hint import HintField
 from onegov.wtfs.fields.municipality import MunicipalityDataUploadField
 
 
 __all__ = (
-    'CsvUploadField',
     'HintField',
     'MunicipalityDataUploadField',
 )