+++ src/onegov.wtfs/onegov/wtfs/tests/test_views_invoice.py
@@ -18,7 +18,7 @@ def test_views_invoices(client):
     upload = client.get('/municipalities').click("Daten importieren")
     upload.form['file'] = Upload(
         'test.csv',
-        "Gemeinde-Nr,Vordefinierte Termine\n1,5.1.2019".encode('utf-8'),
+        "Gemeinde;1;-1;Normal;5.1.2019".encode('cp1252'),
         'text/csv'
     )
     assert "Gemeindedaten importiert." in upload.form.submit().follow()