+++ src/onegov.wtfs/onegov/wtfs/theme/wtfs_theme.py
@@ -4,7 +4,7 @@ from onegov.foundation import BaseTheme
 
 class WtfsTheme(BaseTheme):
     name = 'onegov.wtfs.foundation'
-    version = '1.1.6'
+    version = '1.1.7'
 
     @property
     def pre_imports(self):